Extracted Session class into its own module.
