+++ test/spec-runner.html
@@ -87,6 +87,7 @@
   <script src="../node_modules/material-components-web/dist/material-components-web.js"></script>
 
   <!-- Source files. -->
+  <script type="module" src="/js/session.js"></script>
   <script type="module" src="/js/util.js"></script>
   <script type="module" src="/js/persistence.js"></script>
   <script type="module" src="/js/backend.js"></script>
